# spec/openapi/polaris.v1.yaml
openapi: 3.1.0
info:
  title: Polaris API
  version: 1.0.0
  summary: Machine-readable contract for the Polaris HTTP gateway.
  description: |
    This file is the CI-checkable public HTTP contract for Polaris.
    docs/API_REFERENCE.md remains the human-readable companion.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://gateway.example.com
    description: Production deployment (replace with your own host)
...
security:
  - bearerAuth: []
paths:
...
  /v1/responses:
    post: {operationId: createResponse, tags: [gateway], requestBody: {$ref: "#/components/requestBodies/JSON"}, responses: {"200": {$ref: "#/components/responses/JSONOK"}, default: {$ref: "#/components/responses/Error"}}}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Polaris API key presented as a Bearer token:
        `Authorization: Bearer <key>`. Keys are provisioned through the control
        plane or static config; the gateway persists only their SHA-256 hash.
        MCP endpoints may additionally accept OAuth 2.1 bearer tokens when
        `mcp.auth` includes `oauth`.
...
  requestBodies:
    JSON:
      required: true
      content:
        application/json:
          schema: {type: object}
...
  responses:
    JSONOK:
      description: JSON response
      content:
        application/json:
          schema: {type: object}
    Error:
      description: Polaris error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [message, type]
          additionalProperties: false
          properties:
            message: {type: string}
            type: {type: string}
            code: {type: string}
            param: {type: string}
...